bluesky_scraper/pyproject.toml
[tool.poetry]
name = "scraper-bluesky"
version = "0.1.0"
description = "Scraper of bluesky"
authors = ["skip"]
readme = "README.md"

packages = [
    { include = "scraper_bluesky" },
]

[tool.poetry.scripts]
scraper-bluesky = "scraper_bluesky.__main__:main"


[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.49.0"
opentelemetry-instrumentation = "^0.49b2"
opentelemetry-distro = "^0.49b2"
opentelemetry-exporter-otlp = "^1.28.2"
opentelemetry-instrumentation-asyncio = "^0.49b2"
opentelemetry-instrumentation-dbapi = "^0.49b2"
opentelemetry-instrumentation-logging = "^0.49b2"
opentelemetry-instrumentation-sqlite3 = "^0.49b2"
opentelemetry-instrumentation-threading = "^0.49b2"
opentelemetry-instrumentation-urllib = "^0.49b2"
opentelemetry-instrumentation-wsgi = "^0.49b2"
opentelemetry-instrumentation-grpc = "^0.49b2"
opentelemetry-instrumentation-requests = "^0.49b2"
opentelemetry-instrumentation-urllib3 = "^0.49b2"
opentelemetry-api = "^1.28.2"
opentelemetry-processor-baggage = "^0.49b2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pip-audit = "^2.7.3"
bandit = "^1.7.10"
mypy = "^1.13.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
